+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * add a readonly input field under email for showing the email domain*
 * Add "custom_users" key to config file, to be used with various scripts
 * Add "project managers" that can request admins to add/remove project members
+* Check that user being removed is not owner or last member of a project
+* Add email sent to a project owner when their project is activated
 * Separate global overview from individual group view in admin group management
 * Allow accents in user's first and last names